+++ test/test.webserver.js
@@ -884,7 +884,7 @@ describe('Webserver', () => {
         done();
       });
 
-      it('should retrieve the render file', (done) => {
+      it('should retrieve the render file and set content type and charset', (done) => {
         const body = {
           data : {
             firstname : 'John',
@@ -899,8 +899,31 @@ describe('Webserver', () => {
           assert.strictEqual(err, null);
           get.concat(getBody(4000, `/render/${data.data.renderId}`, 'GET'), (err, res, data) => {
             assert.strictEqual(err, null);
+            assert.strictEqual(res.headers['content-type'], 'text/html; charset=UTF-8');
             assert.strictEqual(res.headers['content-disposition'], 'filename="renderedReport.html"');
             assert.strictEqual(data.toString(), '<!DOCTYPE html> <html> <p>I\'m a Carbone template !</p> <p>I AM John Doe</p> </html> ');
+            done();
+          });
+        });
+      });
+
+      it('should retrieve the render file and set content type and no charset for PDF', (done) => {
+        const body = {
+          data : {
+            firstname : 'John',
+            lastname  : 'Doe'
+          },
+          reportName : 'renderedReport',
+          complement : {},
+          enum       : {},
+          convertTo  : 'pdf'
+        };
+        get.concat(getBody(4000, `/render/${templateId}`, 'POST', body), (err, res, data) => {
+          assert.strictEqual(err, null);
+          get.concat(getBody(4000, `/render/${data.data.renderId}`, 'GET'), (err, res) => {
+            assert.strictEqual(err, null);
+            assert.strictEqual(res.headers['content-type'], 'application/pdf');
+            assert.strictEqual(res.headers['content-disposition'], 'filename="renderedReport.pdf"');
             done();
           });
         });